requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==3.2.23
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
geographiclib==2.0
geopy==2.3.0
idna==3.10
numpy==2.1.2
pandas==2.2.3
polyline==1.4.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
requests==2.32.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==1.26.20
